backend/bayfieldopen/requirements/base.txt
django==3.1.6
requests
gunicorn
djangorestframework
